Fix #8: Add observable collection
Add wrapper around std::unordered_set to implement insertion and removal
of items from a collection. The collection inherits from a value, and
can be subscribed to as usual to yield notifications on assignment. The
collection also provides subscribe_changes which notifies observers on
successful insertion into and deletion from the collection